#726 Updated Trends model to match the interface

+++ Tweetinvi.Logic/Model/Trend.cs
@@ -17,7 +17,7 @@ namespace Tweetinvi.Logic.Model
         [JsonProperty("promoted_content")]
         public string PromotedContent { get; set; }
 
-        [JsonProperty("tweet_volume")] 
-        public int? TweetVolume { get; set; }
+        [JsonProperty("tweet_volume")]
+        public long? TweetVolume { get; set; }
     }
 }
